refactor: Split ApproveWcMethodScreen into smaller widgets
